src/pyproject.toml
[tool.poetry]
name = "semantic-model-generator"
version = "0.1.33"
description = "Generate a Semantic Model from your Snowflake tables"
authors = ["Jonathan Hilgart <jonathan.hilgart@snowflake.com>", "Nipun Sehrawat <nipun.sehrawat@snowflake.com>", "Renee Huang <renee.huang@snowflake.com>", "Nicole Limtiaco <nicole.limtiaco@snowflake.com>"]
license = "Apache Software License; BSD License"
readme = "admin_apps/README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.12"
pandas = "^2.0.1"
loguru = "^0.7.2"
snowflake-connector-python = { extras = ["secure-local-storage", "pandas"], version = "^3.11.0" }
protobuf = "5.26.1"
pydantic = "^2.6.4"
PyYAML = "^6.0.1"
"ruamel.yaml" = "^0.18.6"
jsonargparse = "^4.27.7"
tqdm = "^4.66.2"
pyarrow = ">=10.0.1,<10.1.0"
sqlglot = "^23.17.0"
strictyaml = "^1.7.3"
streamlit = "^1.37.1"
streamlit-extras = "^0.4.3"
streamlit-monaco = "^0.1.3"
numpy = "^1.24.4"
python-dotenv = "^1.0.1"
urllib3 = "^1.26.19"
requests = "^2.32.3"

# Optional dependencies for functionality such as partner semantic model support.
looker-sdk = { version = "^24.14.0", optional = true }

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest = "^8.1.1"
types-pyyaml = "^6.0.12.20240311"
types-protobuf = "^4.24.0.20240311"
pip-licenses = "^4.4.0"
grpcio-tools = "1.64.1"

[tool.poetry.extras]
looker = ["looker-sdk"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
